Vectorize remaining elements of HTTP/1.1 Request
"Vectorizes" parsing of HTTP/1.1 method and version, improves upon
vectorization strategy for target and headers. Note that I might be at a
local minimum here, as the current strategy uses many branches to find
the correct enum value. There may be a more efficient method of
retrieving the enum value via math of some sort, though that is yet to
be explored, and I really need to move on to something else in the
stack.