Move to multiple listener approach
Having a single listener is too slow for handling many requests.
Additionally, the overhead of inter-thread communication and the mutex
synchronization required by it is too high for what we are trying to
accomplish. Instead, use multiple listener threads that handle request
parsing on the spot. Routing overhead should also be minimal, and once
we determine the correct handler, we should be able to kick off an async
runtime specific task to handle processing from there. This means the
listener thread is free to handle more incoming requests from that point
on forward, regardless of how slow the handler is.